refactor: add formCard slice
