+++ app/src/store/index.ts
@@ -2,13 +2,19 @@ import { configureStore } from '@reduxjs/toolkit';
 import searchReducer from './search';
 import searchResultReducer from './searchResult';
 import modalReducer from './modalSlice';
+import formCardsReducer from './formCardsSlice';
 
 const store = configureStore({
   reducer: {
+    formCards: formCardsReducer,
     modal: modalReducer,
     search: searchReducer,
     searchResult: searchResultReducer,
   },
+  middleware: (getDefaultMiddleware) =>
+    getDefaultMiddleware({
+      serializableCheck: false,
+    }),
 });
 
 export type RootState = ReturnType<typeof store.getState>;
