+++ app/src/types/types.ts
@@ -56,3 +56,13 @@ export interface LocationResponse {
   type: string;
   dimension: string;
 }
+
+export type FormData = {
+  name: string;
+  date: string;
+  category: string;
+  condition: string;
+  file: FileList;
+  // file: string;
+  checkbox: boolean;
+};
